requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart 